gcp-functions/customer-review-creator/requirements.txt
functions-framework
confluent-kafka
google-cloud-aiplatform
google-cloud-secret-manager
google-cloud
google
pyyaml